# backend/src/KapitelShelf.Api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "KapitelShelf.Api, Version=0.3.1.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search books with a search term.",
        "parameters": [
          {
            "name": "searchterm",
            "in": "query",
            "description": "The search term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/search/suggestions": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search suggestions with a search term.",
        "parameters": [
          {
            "name": "searchterm",
            "in": "query",
            "description": "The search term.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/series": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Fetch all series.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "The page number.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDTOPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDTOPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDTOPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/series/search": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Search series with the series name.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "The series name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/series/search/suggestions": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Search suggestions with the series name.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "The series name.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/series/watchlist": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Get the series watchlists of a user.",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "The id of the user.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeriesWatchlistDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeriesWatchlistDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeriesWatchlistDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/series/{seriesId}": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Get series by its id.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The id of the series to get.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Series"
        ],
        "summary": "Delete a series.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The id of the seriesto delete.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Series"
        ],
        "summary": "Update a series.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The id of the series to update.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated series.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/series/{seriesId}/merge/{targetSeriesId}": {
      "put": {
        "tags": [
          "Series"
        ],
        "summary": "Merge the series into the target series.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The source series id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "targetSeriesId",
            "in": "path",
            "description": "The target series id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/series/{seriesId}/books": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Get books by the series id.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The series id of the books to get.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/series/{seriesId}/iswatched": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Check, if this series is on the watchlist.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The series id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "The user id.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/series/{seriesId}/watch": {
      "put": {
        "tags": [
          "Series"
        ],
        "summary": "Add a series to the watchlist.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The series id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "The user id.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Series"
        ],
        "summary": "Remove a series from the watchlist.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The series id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "The user id.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/series/{seriesId}/update": {
      "put": {
        "tags": [
          "Series"
        ],
        "summary": "Add a series to the watchlist.",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "description": "The series id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the author id.",
            "format": "uuid"
          },
          "firstName": {
            "type": "string",
            "description": "Gets or sets or set the first name.",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Gets or sets the last name.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The Author DTO."
      },
      "BookDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the book id.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Gets or sets the title.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Gets or sets the description.",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "description": "Gets or sets the release date.",
            "format": "date-time",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "description": "Gets or sets the page number.",
            "format": "int32",
            "nullable": true
          },
          "series": {
            "$ref": "#/components/schemas/SeriesDTO"
          },
          "seriesNumber": {
            "type": "integer",
            "description": "Gets or sets the series number.",
            "format": "int32"
          },
          "author": {
            "$ref": "#/components/schemas/AuthorDTO"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDTO"
            },
            "description": "Gets or sets the categories.",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDTO"
            },
            "description": "Gets or sets the tags.",
            "nullable": true
          },
          "cover": {
            "$ref": "#/components/schemas/FileInfoDTO"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDTO"
          }
        },
        "additionalProperties": false,
        "description": "The book dto."
      },
      "BookDTOPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookDTO"
            },
            "description": "Gets or sets the items.",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "description": "Gets or sets the total number of items.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "The paginated result."
      },
      "CategoryDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the category id.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets or sets the name.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The category dto."
      },
...
      "SeriesDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the series id.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets or sets the name.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Gets or sets the creation time.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Gets or sets the update time.",
            "format": "date-time"
          },
          "lastVolume": {
            "$ref": "#/components/schemas/BookDTO"
          },
          "totalBooks": {
            "type": "integer",
            "description": "Gets or sets the total number of books.",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The series dto."
      },
      "SeriesDTOPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeriesDTO"
            },
            "description": "Gets or sets the items.",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "description": "Gets or sets the total number of items.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "The paginated result."
      },
      "SeriesWatchlistDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the id.",
            "format": "uuid"
          },
          "series": {
            "$ref": "#/components/schemas/SeriesDTO"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookDTO"
            },
            "description": "Gets or sets the watchlist items for this series.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The series watchlist dto."
      },
...
    }
  }
}